requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.0
    # via django
asttokens==2.0.5
    # via stack-data
backcall==0.2.0
    # via ipython
black==22.1.0
    # via ipython
click==8.0.3
    # via black
decorator==5.1.1
    # via ipython
django==3.2.12
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-filter
    #   djangorestframework
django-countries==7.2.1
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
executing==0.8.2
    # via stack-data
ipython==8.0.1
    # via -r requirements.in
jedi==0.18.1
    # via ipython
matplotlib-inline==0.1.3
    # via ipython
mypy-extensions==0.4.3
    # via black
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==2.5.0
    # via black
prompt-toolkit==3.0.27
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.11.2
    # via ipython
pytz==2021.3
    # via
    #   django
    #   djangorestframework
six==1.16.0
    # via asttokens
sqlparse==0.4.2
    # via django
stack-data==0.1.4
    # via ipython
tomli==2.0.1
    # via black
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.0.1
    # via black
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.0.3
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
